refactor server-modules into a folder
